% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayman-ensemble.R
\name{hayman_ensemble_v0_predcard_dt}
\alias{hayman_ensemble_v0_predcard_dt}
\title{Prepare a hayman v0 predcard using predcards for the same forecast_date}
\usage{
hayman_ensemble_v0_predcard_dt(
  predcards_dt,
  test_forecast_date,
  allow_other_generation_dates = FALSE,
  allow_other_prediction_dates = FALSE
)
}
\arguments{
\item{predcards_dt}{a \code{predcards_dt}  containing only the predcards that
should be  incorporated in  the ensemble}

\item{test_forecast_date}{a length-1 \code{\link{Date}}-compatible object}
}
\description{
Prepare a hayman v0 predcard using predcards for the same forecast_date
}
\examples{
library("pipeR")
read_historical_ensemble_components_dt("jhu-csse_deaths_incidence_num","epiweek",2L,"2020-04-06","state",200L) \%>>\%
  testtime_ensemble_components_dt("2020-04-06") \%>>\%
  desired_components_predcards_dt_or_feedback(FALSE, c("strawman"), FALSE, c("aardvark")) \%>>\%
  hayman_ensemble_v0_predcard_dt("2020-04-06") \%>>\%
  predcard_dt_to_predcard_tbl() \%>>\%
  {.}
}
