% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen-ensemble.R
\name{impute_quantile_forecasts}
\alias{impute_quantile_forecasts}
\title{Impute missing forecast quantiles using Evan Ray's method:
Mean imputation for missing forecast quantiles and adjust weights so that
a forecaster gets 0 weight for missing forecast and its weight is distributed
equally to forecasters which provide forecasts for that (date;location)}
\usage{
impute_quantile_forecasts(qf)
}
\description{
Treats a forecast with any NAs as completely missing
}
