% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_based_ensemble.R
\name{panterpolate_raw_jump_cdf}
\alias{panterpolate_raw_jump_cdf}
\title{Get jump CDF from quantiles and raw probs, overweighting top prob}
\usage{
panterpolate_raw_jump_cdf(quantiles, probs)
}
\arguments{
\item{quantiles}{vector of quantiles}

\item{probs}{the levels of these quantiles}
}
\description{
Fits a piecewise constant CDF from the quantiles using approxfun, assigning
\code{1-max(probs)} mass to \code{max(quantiles)+epsilon}.
}
