% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-utils.R
\name{predcard_dt_to_predcard_tbl}
\alias{predcard_dt_to_predcard_tbl}
\title{Convert predcard_dt from \code{data.table} to \code{tibble}}
\usage{
predcard_dt_to_predcard_tbl(predcard_dt)
}
\description{
Convert predcard_dt from \code{data.table} to \code{tibble}
}
