% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-utils.R
\name{predcards_dt_to_predcards_tbl}
\alias{predcards_dt_to_predcards_tbl}
\title{Convert predcards_dt from \code{data.table} to \code{tibble}}
\usage{
predcards_dt_to_predcards_tbl(predcards_dt)
}
\description{
Convert predcards_dt from \code{data.table} to \code{tibble}
}
