% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_based_ensemble.R
\name{quantile_ensemble_via_ruled_crps_with_imputation}
\alias{quantile_ensemble_via_ruled_crps_with_imputation}
\title{Get quantile ensemble by S3-or-ruled-approximation-CRPS-based CDF stacking with imputation}
\usage{
quantile_ensemble_via_ruled_crps_with_imputation(
  qarr,
  y,
  tau,
  panterpolate_cdf,
  weights = NULL,
  prerange.constant.extension = 1,
  prerange.proportional.extension = 0.2
)
}
\arguments{
\item{panterpolate_cdf}{a function like \code{\link{approxfun}} to
interpolate and extrapolate CDF values from a discrete set of xs and ys
(quantiles and probs)}
}
\description{
This is intended to have the same interface as, or one similar to,
\code{\link{quantgen::quantile_ensemble}} function.
}
\details{
This function is a work in progress.

See \code{\link{quantgen::quantile_ensemble}} for arguments.
}
