% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayman-ensemble.R
\name{read_backoff_ensemble_components_dt}
\alias{read_backoff_ensemble_components_dt}
\title{Read backoff component predcards/scorecards into a dt}
\usage{
read_backoff_ensemble_components_dt(
  response,
  incidence_period,
  ahead,
  forecast_dates,
  geo_type,
  n_locations,
  shared_naming_scheme = character(0L),
  only_read_forecaster_names = NULL,
  file_basename = "out.RDS",
  repo_root_dirpath = get_repo_root_dirpath(),
  historical_components_dirpath = get_historical_components_dirpath(repo_root_dirpath),
 
    prospective_components_dirpath = get_prospective_components_dirpath(repo_root_dirpath),
  hub_components_dirpath = get_hub_components_dirpath(repo_root_dirpath),
  historical_naming_scheme = shared_naming_scheme,
  prospective_naming_scheme = shared_naming_scheme,
  hub_naming_scheme = shared_naming_scheme
)
}
\description{
Do not use with untrusted argument values, as this may allow reading of files
from  arbitrary locations.  All character  arguments are  meant to  be actual
values  rather  than  bash  globs. \code{forecast_dates}  can  hold  multiple
values, while most others are expected to be single values.
}
\details{
Here "backoff" refers  to using historical component data if  available for a
given forecast_date and forecaster_name,  otherwise the prospective component
data if available.
}
\examples{
read_backoff_ensemble_components_dt("jhu-csse_deaths_incidence_num","epiweek",2L,"2020-06-22","state",200L,
                                               c("aardvark"="aardvark_vanilla","aardvark_vanilla","poiszero"))

read_backoff_ensemble_components_dt("usafacts_deaths_incidence_num","epiweek",2L,"2020-06-22","county",200L,
                                               c("aardvark"="aardvark_vanilla","aardvark_vanilla","poiszero"))

}
