% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayman-ensemble.R
\name{testtime_ensemble_components_dt}
\alias{testtime_ensemble_components_dt}
\title{Filter predcards_dt to rows for the given test_forecast_date}
\usage{
testtime_ensemble_components_dt(predcards_dt, test_forecast_date)
}
\description{
Filter predcards_dt to rows for the given test_forecast_date
}
\examples{
library("pipeR")
read_backoff_ensemble_components_dt("jhu-csse_deaths_incidence_num","epiweek",2L,"2020-04-06","state",200L,c("aardvark"="aardvark_vanilla","stackedLasso"="stackedLasso_customized"),c("strawman","poiszero","aardvark_vanilla","stackedLasso_customized")) \%>>\%
  testtime_ensemble_components_dt("2020-04-06") \%>>\%
  {.}

}
